# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "version" : "0.0.1",
    "title" : "soeofficer",
    "termsOfService" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "contact" : {
      "name" : "Nordkern ApS",
      "url" : "https://nordkern.dk",
      "email" : "kontakt@nordkern.dk"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/officer" : {
      "post" : {
        "tags" : [ "officer" ],
        "summary" : "Add an officer to the system",
        "description" : "",
        "operationId" : "createOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "appointedNumber",
          "in" : "query",
          "description" : "The appointed number of the officer",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "dodabNumber",
          "in" : "query",
          "description" : "The officer's Dodab number",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "appointedDate",
          "in" : "query",
          "description" : "The date from which the officer title is valid",
          "required" : true,
          "type" : "date"
        }, {
          "name" : "appointedUntil",
          "in" : "query",
          "description" : "The date from which the officer title is invalid",
          "required" : true,
          "type" : "date"
        }, {
          "name" : "terminationCause",
          "in" : "query",
          "description" : "The cause of termination",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "person",
          "in" : "query",
          "required" : true,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Officer"
            }
          }
        }
      }
    },
...
    "/officer/search" : {
      "post" : {
        "tags" : [ "officer" ],
        "summary" : "Get all officers matching the supplied criteria",
        "description" : "",
        "operationId" : "getPersonsMatchingCriteria",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "givenName",
          "in" : "query",
          "description" : "The given name of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "surname",
          "in" : "query",
          "description" : "The surname of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateOfBirth",
          "in" : "query",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "gender",
          "in" : "query",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "appointedDate",
          "in" : "query",
          "description" : "The date from which the officer title is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "rankID",
          "in" : "query",
          "description" : "The ID of the rank to be used as a criteria",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "dodabNumber",
          "in" : "query",
          "description" : "The officer's Dodab number",
          "required" : false,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Officer"
              }
            }
          }
        }
      }
    },
...
    "/officer/{id}" : {
      "get" : {
        "tags" : [ "officer" ],
        "summary" : "Get specific officer from id",
        "description" : "",
        "operationId" : "readOfficer",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Officer"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "officer" ],
        "summary" : "Update an existing officer of the system",
        "description" : "",
        "operationId" : "updateOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "appointedNumber",
          "in" : "query",
          "description" : "The appointed number of the officer",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "dodabNumber",
          "in" : "query",
          "description" : "The officer's Dodab number",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "appointedDate",
          "in" : "query",
          "description" : "The date from which the officer title is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "appointedUntil",
          "in" : "query",
          "description" : "The date from which the officer title is invalid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "terminationCause",
          "in" : "query",
          "description" : "The cause of termination",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "person",
          "in" : "query",
          "required" : false,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Officer"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "officer" ],
        "summary" : "Delete an officer from the system",
        "description" : "",
        "operationId" : "deleteOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/officer/{id}/limited" : {
      "get" : {
        "tags" : [ "officer" ],
        "summary" : "Get specific officer from id (limited)",
        "description" : "",
        "operationId" : "readOfficerLimited",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Officer"
            }
          }
        }
      }
    },
    "/officer/{id}/promote" : {
      "post" : {
        "tags" : [ "officer" ],
        "summary" : "Promote an officer",
        "description" : "",
        "operationId" : "promoteOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Promotion"
            }
          }
        }
      }
    },
    "/officer/{id}/promotion/all" : {
      "get" : {
        "tags" : [ "officer" ],
        "summary" : "Get all promotions belonging to an officer",
        "description" : "",
        "operationId" : "getPromotions",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Promotion"
              }
            }
          }
        }
      }
    },
...
    "/person/search" : {
      "post" : {
        "tags" : [ "person" ],
        "summary" : "Get all persons in the system matching the supplied criteria",
        "description" : "",
        "operationId" : "getPersonsMatchingCriteria",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "givenName",
          "in" : "query",
          "description" : "The given name of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "surname",
          "in" : "query",
          "description" : "The surname of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateOfBirth",
          "in" : "query",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "gender",
          "in" : "query",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Person"
              }
            }
          }
        }
      }
    },
...
    "/promotion/{id}" : {
      "put" : {
        "tags" : [ "promotion" ],
        "summary" : "Update a promotion of the system",
        "description" : "",
        "operationId" : "updatePromotion",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "dateOfPromotion",
          "in" : "query",
          "description" : "The date from which the rank is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "rank",
          "in" : "query",
          "description" : "The date from which the rank is invalid",
          "required" : false,
          "type" : "object"
        }, {
          "name" : "officer",
          "in" : "query",
          "description" : "The id of the officer possessing the rank",
          "required" : false,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Promotion"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "promotion" ],
        "summary" : "Delete a promotion from the system",
        "description" : "",
        "operationId" : "deleteOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/user/search" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Get all users in the system matching the supplied criteria",
        "description" : "",
        "operationId" : "getPersonsMatchingCriteria",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "The username of the user",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "email",
          "in" : "query",
          "description" : "The email of the user",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Officer" : {
      "type" : "object",
      "required" : [ "appointedNumber", "dodabNumber", "id", "person", "terminationCause" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the officer"
        },
        "appointedNumber" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 12,
          "description" : "The appointed number of the officer"
        },
        "dodabNumber" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 123,
          "description" : "The officer's Dodab number"
        },
        "appointedDate" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the officer title is valid",
          "readOnly" : true
        },
        "appointedUntil" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the officer title is invalid"
        },
        "terminationCause" : {
          "type" : "string",
          "example" : "Accidental_death",
          "description" : "The cause of termination",
          "enum" : [ "Terminated", "Killed_in_action", "Accidental_death", "Transferred", "Other" ]
        },
        "person" : {
          "description" : "The person associated with the officer",
          "$ref" : "#/definitions/Person"
        }
      }
    },
...
    "Person" : {
      "type" : "object",
      "required" : [ "dateOfBirth", "gender", "givenName", "id", "surname" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the person"
        },
        "givenName" : {
          "type" : "string",
          "example" : "John",
          "description" : "The given name of the person"
        },
        "surname" : {
          "type" : "string",
          "example" : "Doe",
          "description" : "The surname of the person"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY"
        },
        "gender" : {
          "type" : "string",
          "example" : "Mand",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "enum" : [ "Male", "Female", "Unknown" ]
        },
        "dateOfDeath" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date of death. Format is: DD/MM/YYYY"
        }
      }
    },
...
    "Promotion" : {
      "type" : "object",
      "required" : [ "dateOfPromotion", "id", "rank" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the promotion"
        },
        "dateOfPromotion" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "1",
          "description" : "The date from which the rank is valid"
        },
        "rank" : {
          "example" : "1",
          "description" : "The date from which the rank is invalid",
          "$ref" : "#/definitions/Rank"
        }
      }
    },
    "Rank" : {
      "type" : "object",
      "required" : [ "id", "natoCode", "rankName", "rankNameEn", "rankValidFrom", "rankValidUntil" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the rank"
        },
        "natoCode" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1234,
          "description" : "The nato code of the rank"
        },
        "rankName" : {
          "type" : "string",
          "example" : "Officer",
          "description" : "The name of the rank"
        },
        "rankNameEn" : {
          "type" : "string",
          "example" : "Officer",
          "description" : "The english name of the rank"
        },
        "rankValidFrom" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the rank is valid"
        },
        "rankValidUntil" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the rank is invalid"
        }
      }
    },
...
    "User" : {
      "type" : "object",
      "required" : [ "email", "from", "id", "role", "until", "username" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the user"
        },
        "username" : {
          "type" : "string",
          "example" : "john_doe",
          "description" : "The unique username of the user"
        },
        "email" : {
          "type" : "string",
          "example" : "john@doe.com",
          "description" : "The unique email of the user"
        },
        "from" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the user is valid. Format is: DD/MM/YYYY"
        },
        "until" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "31/12/9999",
          "description" : "The date on which to user is invalid. Format is: DD/MM/YYYY"
        },
        "role" : {
          "type" : "string",
          "example" : "administrator",
          "description" : "The system-role of the user",
          "enum" : [ "contributor", "administrator", "read", "privileged_read" ]
        }
      }
    }
  }
}